changed split-phase semantics of slave rx/tx
If xxxRequested returns TRUE, it means "I already responded, so leave
the interrupts enabled." The event handler MUST do xxx before it
returns. If it doesn't, the interrupt will be hit again as soon as the
reti instruction is hit.

Returning FALSE will disable the relevant interrupt until the client
can finish dealing with it.

It should probably also disable state-change interrupts temporarily
(otherwise, the behavior at the last byte is potentially weird).