hack for stop/rx interrupt priorities
On the cc430, the stop interrupt has higher priority than the rx
interrupt.  So, when we get a stop interrupt we have to deal with the
data that might be outstanding. Note that if the application tries to
delay reading the last byte from the RXBUF, there's no real guarantee
that it will be around when they try to read it. Hopefully, the logic
at the application level notices that the stop condition has happened
and doesn't try it.