resolved interrupt priority issue.
There is a very short window between the final RX interrupt of a
transaction and the following STOP or repeated START interrupt.  On
the cc430, STOP/START have higher priority, so we tended to see them
happen first, which messed up the upper-layer's logic.  We manually
check to see whether an RX interrupt *should* be ready (RXIFG set, RX
interrupt enabled) when we get a stop or start condition, and handle
that interrupt first.