added force-reset option. hack for bsl entry
memory programming guide specifies TEST needs 2 RE transitions while
RSTn is low. Doing 3 seems to greatly improve reliability.

Found that if there was a running application, we needed to do a
standard reset before the bsl entry. If we had erased the memory, then
we need to NOT do a standard reset before bsl entry. Adding the third
rising edge (which was the result of a complete guess) seems to make
the bsl entry work correctly regardless of what state the memory is
in.